smartfarmassistant/smartfarmassistant/requirements.txt
Django
djangorestframework
psycopg2-binary
scikit-learn
paho-mqtt